requirements.txt
Pillow
opencv-python
librosa
termcolor
inquirer 
